Refactor from `enum` subpackage to `particle_tracker` module
