--- docs/enum/materials.rst
@@ -1,9 +1,0 @@
-.. _materials:
-
-******************************************************************
-Material Enumerations (`plasmapy.enum.materials`)
-******************************************************************
-
-.. currentmodule:: plasmapy.enum.materials
-
-.. automodapi:: plasmapy.enum.materials
