--- src/plasmapy/enum/__init__.py
@@ -1,8 +1,0 @@
-"""
-The enum subpackage stores the enumerations of acceptable parameters for
-related functions.
-"""
-
-__all__ = ["StoppingMaterial"]
-
-from materials import StoppingMaterial
