+++ src/plasmapy/simulation/particle_tracker.py
@@ -9,6 +9,7 @@ __all__ = [
     "IntervalSaveRoutine",
     "NoParticlesOnGridsTerminationCondition",
     "ParticleTracker",
+    "StoppingMaterial",
     "TimeElapsedTerminationCondition",
 ]
 
@@ -17,6 +18,7 @@ import sys
 import warnings
 from abc import ABC, abstractmethod
 from collections.abc import Iterable
+from enum import StrEnum, auto
 from pathlib import Path
 from typing import Optional
 
@@ -30,6 +32,87 @@ from plasmapy.particles import Particle, particle_input
 from plasmapy.plasma.grids import AbstractGrid
 from plasmapy.plasma.plasma_base import BasePlasma
 from plasmapy.simulation.particle_integrators import boris_push
+
+
+class StoppingMaterial(StrEnum):
+    """
+    Acceptable materials to be used with the NIST stopping power database.
+    """
+
+    ACETYLENE = auto()
+    ADIPOSE_TISSUE_ICRP = auto()
+    AIR_DRY_NEAR_SEA_LEVEL = auto()
+    ALUMINUM = auto()
+    ALUMINUM_OXIDE = auto()
+    AMORPHOUS_CARBON = auto()
+    ARGON = auto()
+    A_150_TISSUE_EQUIVALENT_PLASTIC = auto()
+    BERYLLIUM = auto()
+    BONE_COMPACT_ICRU = auto()
+    BONE_CORTICAL_ICRP = auto()
+    B_100_BONE_EQUIVALENT_PLASTIC = auto()
+    CALCIUM_FLUORIDE = auto()
+    CARBON_DIOXIDE = auto()
+    CELLULOSE_NITRATE = auto()
+    CERIC_SULFATE_DOSIMETER_SOLUTION = auto()
+    CESIUM_IODIDE = auto()
+    COPPER = auto()
+    C_552_AIR_EQUIVALENT_PLASTIC = auto()
+    ETHYLENE = auto()
+    FERROUS_SULFATE_DOSIMETER_SOLUTION = auto()
+    GADOLINIUM = auto()
+    GERMANIUM = auto()
+    GOLD = auto()
+    GRAPHITE = auto()
+    HELIUM = auto()
+    HYDROGEN = auto()
+    IRON = auto()
+    KAPTON_POLYIMIDE_FILM = auto()
+    KRYPTON = auto()
+    LEAD = auto()
+    LITHIUM_FLUORIDE = auto()
+    LITHIUM_TETRABORATE = auto()
+    M3_WAX = auto()
+    METHANE = auto()
+    MOLYBDENUM = auto()
+    MS20_TISSUE_SUBSTITUTE = auto()
+    MUSCLE_EQUIVALENT_LIQUID_WITHOUT_SUCROSE = auto()
+    MUSCLE_EQUIVALENT_LIQUID_WITH_SUCROSE = auto()
+    MUSCLE_SKELETAL_ICRP = auto()
+    MUSCLE_STRIATED_ICRU = auto()
+    NEON = auto()
+    NITROGEN = auto()
+    NYLON_TYPE_6_AND_TYPE_6_6 = auto()
+    OXYGEN = auto()
+    PARAFFIN_WAX = auto()
+    PHOTOGRAPHIC_EMULSION = auto()
+    PLASTIC_SCINTILLATOR_VINYLTOLUENE_BASED = auto()
+    PLATINUM = auto()
+    POLYCARBONATEMAKROLON_LEXAN = auto()
+    POLYETHYLENE = auto()
+    POLYETHYLENE_TEREPHTHALATE_MYLAR = auto()
+    POLYMETHYL_METHACRALATE_LUCITE_PERSPEX = auto()
+    POLYPROPYLENE = auto()
+    POLYSTYRENE = auto()
+    POLYTETRAFLUOROETHYLENE_TEFLON = auto()
+    POLYVINYL_CHLORIDE = auto()
+    PROPANE = auto()
+    Pyrex_Glass = auto()
+    SILICON = auto()
+    SILICON_DIOXIDE = auto()
+    SILVER = auto()
+    SODIUM_IODIDE = auto()
+    STILBENE = auto()
+    TIN = auto()
+    TISSUE_EQUIVALENT_GAS_METHANE_BASED = auto()
+    TISSUE_EQUIVALENT_GAS_PROPANE_BASED = auto()
+    TITANIUM = auto()
+    TOLUENE = auto()
+    TUNGSTEN = auto()
+    URANIUM = auto()
+    WATER_LIQUID = auto()
+    WATER_VAPOR = auto()
+    XENON = auto()
 
 
 class AbstractTerminationCondition(ABC):
